added System Process Types and const values so that I can manage them a little easier, if they expand

+++ prisma/seed.ts
@@ -7,9 +7,11 @@ import {
   SessionAction,
   SessionResource,
   SessionRole,
+  SYSTEM_PROCESS_MANIFEST,
   TColors,
   TConfigurationKeys,
   TStatusKey,
+  TSystemProcess,
 } from '../lib/types';
 import { addDays, addYears, differenceInDays, endOfDay, startOfDay } from 'date-fns';
 import {
@@ -371,17 +373,6 @@ async function getActiveEventStatus(key: TStatusKey): Promise<{ statusId: number
   });
 
   return result;
-}
-
-function FindRoomID(
-  roomName: string,
-  rooms: { name: string; createdAt: Date; updatedAt: Date; roomId: number; color: string; icon: string | null }[],
-) {
-  const test = rooms.find((room) => {
-    return room.name === roomName;
-  });
-
-  return test?.roomId ?? 0;
 }
 
 function getRandomDescription(): string {
@@ -1208,6 +1199,47 @@ async function createLinkedServer() {
   await prisma.$executeRawUnsafe(`CALL public.insert_avanti_users();`);
 }
 
+async function FindCreateSystemProcess(processKey: TSystemProcess) {
+  const SYSTEM_PROCESS_KEY = SYSTEM_PROCESS_MANIFEST[processKey].key;
+  const SYSTEM_PROCESS_DEFAULT_PARAMETER = SYSTEM_PROCESS_MANIFEST[processKey].defaultParameter;
+
+  const processTag = `${SYSTEM_PROCESS_KEY}_${process.env.DATABASE_NAME ? process.env.DATABASE_NAME : 'unknown'}`.toUpperCase();
+
+  await prisma.systemProcess.upsert({
+    create: { pid: 0, key: processKey, tag: processTag, parameter: SYSTEM_PROCESS_DEFAULT_PARAMETER, createdBy: 0, updatedBy: 0 },
+    update: {},
+    where: { key: processKey },
+    select: { pid: true, tag: true, updatedAt: true },
+  });
+}
+
+export async function saveSystemProcess({
+  pid,
+  parameter,
+  processTag,
+  processKey,
+  userId = 0,
+}: {
+  pid: number;
+  parameter: string;
+  processTag: string;
+  processKey: string;
+  userId?: number;
+}) {
+  try {
+    const process = await prisma.systemProcess.upsert({
+      create: { pid, key: processKey, tag: processTag, parameter: parameter, createdBy: userId, updatedBy: userId },
+      update: { pid, parameter: parameter, updatedBy: userId },
+      where: { key: processKey },
+      select: { pid: true, tag: true, updatedAt: true },
+    });
+
+    return process;
+  } catch (err) {
+    console.error('[Scheduler] Failed to save metadata:', err);
+  }
+}
+
 async function deleteAllData() {
   await prisma.eventRoom.deleteMany();
   await prisma.eventRecipient.deleteMany();
@@ -1241,8 +1273,12 @@ async function deleteAllData() {
 }
 
 async function main() {
-  console.log('Deleting All Data...');
-  await deleteAllData();
+  if (process.env.NEXT_PUBLIC_ENVIRONMENT !== 'production') {
+    console.log('Deleting All Data...');
+    await deleteAllData();
+  }
+
+  FindCreateSystemProcess('ENTRA_SYNC_SCHEDULER');
 
   console.log('Creating System User...');
   const SYSTEM_USER = await prisma.user.upsert({
@@ -1261,19 +1297,24 @@ async function main() {
     },
   });
 
-  /*if (process.env.NEXT_PUBLIC_ENVIRONMENT === 'production') {
-    await prisma.session.deleteMany();
-    await prisma.account.deleteMany();
-    await prisma.user.deleteMany();
-    await prisma.sSOProvider.deleteMany();
-    await prisma.userRole.deleteMany();
-    await prisma.verification.deleteMany();
+  if (process.env.ADMIN_USER_EMAIL) {
+    const ADMIN_USER = await prisma.user.upsert({
+      where: { email: process.env.ADMIN_USER_EMAIL },
+      update: {},
+      create: {
+        name: 'Admin User',
+        email: process.env.ADMIN_USER_EMAIL,
+        emailVerified: false,
+        image: null,
+        externalId: '000',
+        isActive: true,
+        createdBy: 0,
+        updatedBy: 0,
+      },
+    });
+    const adminRole = await FindCreateRole('Admin');
+    const adminUserRole = await FindCreateUserRole(adminRole.roleId, ADMIN_USER.id);
   }
-
-  await prisma.role.deleteMany();
-  await prisma.action.deleteMany();
-  await prisma.resource.deleteMany();
-  await prisma.roleResourceAction.deleteMany();*/
 
   if (process.env.NEXT_PUBLIC_ENVIRONMENT === 'development') {
     console.log('Seeding Random Users...');
@@ -1365,25 +1406,6 @@ async function main() {
     await createLinkedServer();
   }
 
-  if (process.env.ADMIN_USER_EMAIL) {
-    const ADMIN_USER = await prisma.user.upsert({
-      where: { email: process.env.ADMIN_USER_EMAIL },
-      update: {},
-      create: {
-        name: 'Admin User',
-        email: process.env.ADMIN_USER_EMAIL,
-        emailVerified: false,
-        image: null,
-        externalId: '000',
-        isActive: true,
-        createdBy: 0,
-        updatedBy: 0,
-      },
-    });
-    const adminRole = await FindCreateRole('Admin');
-    const adminUserRole = await FindCreateUserRole(adminRole.roleId, ADMIN_USER.id);
-  }
-
   if (process.env.NEXT_PUBLIC_ENVIRONMENT === 'development') {
     console.log('Seeding Random Events...');
     await prisma.event.deleteMany();
@@ -1398,47 +1420,6 @@ async function main() {
     console.log('Seeding Edge Case Multi-Day Events...');
     await CreateEdgeCaseMultiDayEvents(roomList);
   }
-
-  /*
-
-
-  const alice = await prisma.user.upsert({
-    where: { email: "alice@prisma.io" },
-    update: {},
-    create: {
-      email: "alice@prisma.io",
-      name: "Alice",
-      posts: {
-        create: {
-          title: "Check out Prisma with Next.js",
-          content: "https://www.prisma.io/nextjs",
-          published: true,
-        },
-      },
-    },
-  });
-  const bob = await prisma.user.upsert({
-    where: { email: "bob@prisma.io" },
-    update: {},
-    create: {
-      email: "bob@prisma.io",
-      name: "Bob",
-      posts: {
-        create: [
-          {
-            title: "Follow Prisma on Twitter",
-            content: "https://twitter.com/prisma",
-            published: true,
-          },
-          {
-            title: "Follow Nexus on Twitter",
-            content: "https://twitter.com/nexusgql",
-            published: true,
-          },
-        ],
-      },
-    },
-  });*/
 }
 main()
   .then(async () => {
